Change `BFOR` / `}for` to use relative addressing
Previously, for-loops used an absolute address for the backwards
branch at the end of the loop. After this change, the branch for
`BFOR` uses a relative offset. The jump is always backwards, so
offsets are stored as a uint8 (max 255) to be subtracted from PC.

+++ repl2/bootstrap.py
@@ -279,9 +279,11 @@ class Compiler:
       return pos + 1
     if w == '}for':                     # }for
       self.append_byte(BFOR)
-      addr = self.vm.T
+      offset = self.DP - self.vm.T
+      if offset > 255:
+        raise Exception("`for{` to `}for` distance is > 255")
       self.vm.drop()
-      self.append_halfword(addr)
+      self.append_byte(offset)
       self.nest_for -= 1
       if self.nest_for < 0:
         raise Exception("`}for` without matching `for{`")
